Github actions implemnetation for github documentation sync. Optimisation of Google Tag Manager. Minor layout changes. Refractor settings for simplicity and controls. Adjustments to documentations.
