+++ components/navigation/footer.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import Image from "next/image";
-import { Company } from "@/settings/config";
+import { Company } from "@/lib/meta";
 
 export function Footer() {
   return (
