+++ contents/docs/introduction/changelog/index.mdx
@@ -3,8 +3,6 @@ title: Changelog
 description: This section provides an overview of Changelog.
 ---
 
-![Alt text](/images/og-image.png "Documents")
-
 Lorem ipsum dolor sit amet consectetur adipisicing elit. Numquam iste dolorum tempore consectetur explicabo tempora provident quia maxime, molestiae, facilis aperiam et, error illum vel ullam? Quis architecto dolore ullam
 
 <Note title="Remember" type="danger">
