+++ contents/docs/introduction/index.mdx
@@ -3,6 +3,8 @@ title: Introduction
 description: This section provides an overview of how to get started with the Documents, Next.js Document Starter Kit.
 ---
 
+![Banner](/images/banner.png "Documents")
+
 ## Documents
 
 Welcome to **Documents**, a powerful and flexible Next.js documentation starter kit designed to simplify the process of creating high-quality, comprehensive 
@@ -19,6 +21,12 @@ insights into how to structure and present your content effectively.
     title="Installation"
     description="Get started with Documents using our quick start installation guide to get your project started."
     href="/docs/introduction/installation"
+  />
+  <Card
+    subtitle="Setup"
+    title="Settings"
+    description="Setting up your Dcouments projects layout, links and search engine optimisation."
+    href="/docs/introduction/setup"
   />
   <Card
     subtitle="Support"
