+++ settings/settings.ts
@@ -1,0 +1,23 @@
+export const url = "https://rubix-documents.vercel.app";
+export const siteicon = "/icon.png";
+
+export const gtm = "GTM-XXXXXXX";
+export const gtmconnected = true;
+
+export const sitename = "Documents";
+export const description = "This comprehensive documentation template, crafted with Next.js and available as open-source, delivers a sleek and responsive design, tailored to meet all your project documentation requirements.";
+export const keywords = ["Next.js", "documentation", "Rubix Studios", "open source", "SEO"];
+export const urlimage = "/images/og-image.png";
+export const imagealt = "Rubix Studios Documentation";
+export const twitterhandle = "@RubixStudios";
+
+export const companyname = "Rubix Studios";
+export const companylink = "https://www.rubixstudios.com.au";
+
+export const branding = true;
+export const rightsidebar = true;
+export const feedbackedit = true;
+export const tableofcontent = true;
+export const totopscroll = true;
+
+export const loadfromgithub = false;
